finishing editing cabin

+++ src/services/apiCabins.js
@@ -1,44 +1,59 @@
-import supabase from "./supabase";
-import { supabaseUrl } from "./supabase";
+import supabase, { supabaseUrl } from "./supabase";
+
 export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.error("Error fetching cabins:", error);
-    throw new Error(error.message);
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
   }
+
   return data;
 }
 
-export async function addCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabins-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabins-images/${imageName}`;
 
-  //1-create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
   if (error) {
-    console.error("Error inserting cabin:", error);
-    throw new Error(error.message);
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
-  //2-upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabins-images")
     .upload(imageName, newCabin.image);
-  console.log(newCabin.image);
-  //3- delete cabin if the is an error  uploading image
+
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
-      "Error cabin image  could not be  uploaded and   the cabin was not created"
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
+
   return data;
 }
 
@@ -46,8 +61,9 @@ export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
 
   if (error) {
-    console.error("Error deleting cabin:", error);
-    throw new Error(error.message);
+    console.error(error);
+    throw new Error("Cabin could not be deleted");
   }
+
   return data;
 }
